Added LDAP authentication, added option for no authentication, restructured files in security, enhanced session management
