+++ requirements.txt
@@ -1,56 +1,13 @@
-aiohappyeyeballs==2.4.4
 aiohttp==3.11.11
-aiosignal==1.3.2
-annotated-types==0.7.0
-anyio==4.8.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-async-timeout==5.0.1
-attrs==25.1.0
-cachetools==5.5.1
-certifi==2025.1.31
-cffi==1.17.1
-charset-normalizer==3.4.1
-click==8.1.8
 croniter==6.0.0
-durationpy==0.9
-ecdsa==0.19.0
-exceptiongroup==1.2.2
 fastapi==0.115.8
-frozenlist==1.5.0
-google-auth==2.38.0
-greenlet==3.1.1
-h11==0.14.0
-idna==3.10
 kubernetes==32.0.0
 minio==7.2.15
-multidict==6.1.0
-nats-py==2.8.0
-oauthlib==3.2.2
 passlib==1.7.4
-propcache==0.2.1
-pyasn1==0.6.1
-pyasn1_modules==0.4.1
-pycparser==2.22
-pycryptodome==3.21.0
 pydantic==2.10.6
-pydantic_core==2.27.2
-python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
-python-jose==3.3.0
-python-multipart==0.0.19
-pytz==2025.1
-PyYAML==6.0.2
 requests==2.32.3
-requests-oauthlib==2.0.0
-rsa==4.9
-six==1.17.0
-sniffio==1.3.1
 SQLAlchemy==2.0.37
 starlette==0.45.3
-typing_extensions==4.12.2
-urllib3==2.3.0
 uvicorn==0.34.0
-websocket-client==1.8.0
-wsproto==1.2.0
-yarl==1.18.3
+ldap3~=2.9.1
