+++ src/api/common/routers/health.py
@@ -7,8 +7,8 @@ from core.config import ConfigHelper
 from utils.commons import route_description
 from utils.handle_exceptions import handle_exceptions_endpoint
 
-from security.service.helpers.rate_limiter import LimiterRequests
-from security.service.helpers.rate_limiter import RateLimiter
+from security.helpers.rate_limiter import LimiterRequests
+from security.helpers.rate_limiter import RateLimiter
 
 from api.common.response_model.failed_request import FailedRequest
 from api.common.response_model.successful_request import SuccessfulRequest
