+++ src/security/routers/authentication.py
@@ -1,21 +1,14 @@
 from typing import Annotated
-from fastapi import APIRouter
-# from requests import Response
-# from sqlalchemy.ext.asyncio import AsyncSession
-
-# from fastapi.security import OAuth2PasswordRequestForm
-from security.helpers.request import OAuth2PasswordAndRefreshRequestForm
+from fastapi import APIRouter, Depends
+from security.schemas.request.OAuth2_request import OAuth2PasswordAndRefreshRequestForm
 from typing import Union
 from security.controllers.authentication import Authentication
-from security.model.model import Token, TokenRefresh
-from security.service.helpers.rate_limiter import LimiterRequests, RateLimiter
-from security.service.helpers.users import *
+from security.schemas.response.token import TokenRefresh, Token
+from security.helpers.rate_limiter import LimiterRequests, RateLimiter
 
 from utils.commons import route_description
 
 from core.config import ConfigHelper
-
-# from utils.handle_exceptions import handle_exceptions_async_method
 
 config = ConfigHelper()
 router = APIRouter()
@@ -42,12 +35,13 @@ route = '/token'
              response_model=Union[Token, TokenRefresh])
 # async def login_for_access_token(form_data: Annotated[OAuth2PasswordRequestForm, Depends()],
 #                                  db: Session = Depends(get_db)):
-async def login_for_access_token(form_data: Annotated[OAuth2PasswordAndRefreshRequestForm, Depends()],
-                                 db: Session = Depends(get_db)):
+# async def login_for_access_token(form_data: Annotated[OAuth2PasswordAndRefreshRequestForm, Depends()],
+#                                 db: Session = Depends(get_db)):
+async def login_for_access_token(form_data: Annotated[OAuth2PasswordAndRefreshRequestForm, Depends()]):
     if form_data.grant_type == "refresh_token":
-        return await auth.refresh_login(form_data.refresh_token, db)
+        return await auth.refresh_login(form_data.refresh_token)
     else:
-        return await auth.login(form_data.username, form_data.password, db)
+        return await auth.login(form_data.username, form_data.password)
 
 # limiter_tr = endpoint_limiter.get_limiter_cust('token_refresh')
 # route = '/token-refresh'
