+++ src/security/schemas/response/token.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+
+
+class TokenRefresh(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+    refresh_token: str
