+++ src/security/schemas/runtime/token.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class TokenSession(BaseModel):
+    sub: str | None = None
+    username: str | None = None
+    auth_type: str | None = None
